after change- at

+++ BM-Client/gui/client/controllers/RestListFormController.java
@@ -91,7 +91,7 @@ public class RestListFormController extends Controller implements Initializable 
 		if(supplier!=null)
 		{
 			String address=table.getSelectionModel().getSelectedItem().getAddress();
-			Message msg= new Message(MessageType.get_Dishes,(Integer)table.getSelectionModel().getSelectedItem().getRestCode());
+			Message msg= new Message(MessageType.get_Dishes,table.getSelectionModel().getSelectedItem().getRestCode());
 			ClientUI.chat.accept(msg);
 			System.out.println(dishes.get(0).getDishName());
